# Conflicts:
#	src/app/modules/articles/components/list-articles/list-articles.component.ts
#	src/app/modules/stock-items/components/list-stock-items/list-stock-items.component.ts

+++ src/app/modules/articles/components/list-articles/list-articles.component.ts
@@ -9,22 +9,8 @@ import {formatNumber} from '@angular/common';
 import {MatTableDataSource} from "@angular/material/table";
 import {MatPaginator} from "@angular/material/paginator";
 import {MatSort} from "@angular/material/sort";
-export interface UserData {
-  id: string;
-  name: string;
-  progress: string;
-  color: string;
-}
+import {ConfirmationDialogComponent} from "../../../shared/components/confirmation-dialog/confirmation-dialog.component";
 
-/** Constants used to fill up our data base. */
-const COLORS: string[] = [
-  'maroon', 'red', 'orange', 'yellow', 'olive', 'green', 'purple', 'fuchsia', 'lime', 'teal',
-  'aqua', 'blue', 'navy', 'black', 'gray'
-];
-const NAMES: string[] = [
-  'Maia', 'Asher', 'Olivia', 'Atticus', 'Amelia', 'Jack', 'Charlotte', 'Theodore', 'Isla', 'Oliver',
-  'Isabella', 'Jasper', 'Cora', 'Levi', 'Violet', 'Arthur', 'Mia', 'Thomas', 'Elizabeth'
-];
 
 @Component({
   selector: 'app-list-articles',
@@ -88,12 +74,19 @@ export class ListArticlesComponent implements OnInit {
   }
 
   deleteArticle(article: Article): void {
-    this.articlesService.deleteArticle(article).subscribe(
-      res => this.ngOnInit()
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent);
+    dialogRef.componentInstance.confirmMessage = "Are you sure you want to delete?";
+    dialogRef.afterClosed().subscribe(result => {
+        if(result){
+          this.articlesService.deleteArticle(article).subscribe(
+            res => this.ngOnInit()
+          );
+        }
+      }
     );
   }
 
-  addCard(article: Article) {
+addCard(article: Article) {
     this.router.navigateByUrl('stockitems/create/' + article.id);
   }
 
@@ -113,20 +106,4 @@ export class ListArticlesComponent implements OnInit {
     this.router.navigateByUrl('stockitems/list/id/' + id + '/searchBy/article');
   }
 
-}
-
-
-
-
-/** Builds and returns a new User. */
-function createNewUser(id: number): UserData {
-  const name = NAMES[Math.round(Math.random() * (NAMES.length - 1))] + ' ' +
-    NAMES[Math.round(Math.random() * (NAMES.length - 1))].charAt(0) + '.';
-
-  return {
-    id: id.toString(),
-    name: name,
-    progress: Math.round(Math.random() * 100).toString(),
-    color: COLORS[Math.round(Math.random() * (COLORS.length - 1))]
-  };
 }
