+++ src/app/modules/users/components/list-roles/list-roles.component.ts
@@ -4,6 +4,7 @@ import {Role} from '../../entities/role';
 import {UpdateRoleDialogComponent} from '../update-role-dialog/update-role-dialog.component';
 import {MatDialog} from "@angular/material/dialog";
 import {Router} from '@angular/router';
+import {ConfirmationDialogComponent} from "../../../shared/components/confirmation-dialog/confirmation-dialog.component";
 import {MatTableDataSource} from "@angular/material/table";
 import {UserData} from "../../../articles/components/material-example/material-example.component";
 import {MatSort} from "@angular/material/sort";
@@ -53,8 +54,15 @@ export class ListRolesComponent implements OnInit {
   }
 
   deleteRole(role: Role): void {
-    this.usersService.deleteRole(role).subscribe(
-      res => this.ngOnInit()
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent);
+    dialogRef.componentInstance.confirmMessage = "Are you sure you want to delete?";
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        this.usersService.deleteRole(role).subscribe(
+          res => this.ngOnInit()
+        );
+      }
+      }
     );
   }
 
@@ -66,4 +74,5 @@ export class ListRolesComponent implements OnInit {
       this.dataSource.paginator.firstPage();
     }
   }
+
 }
